+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/controller/TigerWebUiControllerTest.java
@@ -18,7 +18,6 @@ package de.gematik.test.tiger.proxy.controller;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.await;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.core.StringContains.containsString;