+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/servers/ExternalJarServer.java
@@ -51,7 +51,7 @@ public class ExternalJarServer extends AbstractExternalTigerServer {
   @Builder
   public ExternalJarServer(
       TigerTestEnvMgr tigerTestEnvMgr, String serverId, CfgServer configuration) {
-    super(determineHostname(configuration, serverId), serverId, configuration, tigerTestEnvMgr);
+    super(serverId, configuration, tigerTestEnvMgr);
   }
 
   @Override